+++ tests/test_cli.py
@@ -19,6 +19,7 @@ ENTRY_POINTS = [
     ("pytoolbox.pytime", "time_cli"),
     ("pytoolbox.pyssh", "ssh_management"),
     ("pytoolbox.pynet", "net_cli"),
+    ("pytoolbox.pycalc", "calc_cli"),
     ("pytoolbox.pymd2pdf", "pymd2pdf_cli"),
     ("pytoolbox.pymd2html", "md2html_cli"),
     ("pytoolbox.pydocx2pdf", "docx2pdf_cli"),